requirements.txt
numpy
pandas
scikit-learn

pyarrow
pandas-gbq

spacy
sentence-transformers
rapidfuzz

tqdm

google-cloud-storage	
google-cloud-bigquery
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
google-cloud-bigquery
